requirements.txt
lxml==4.6.2
numpy==1.20.1
Pillow==8.1.0
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
